Renamed SearchPluginsWindow to RemoteSearchPluginsWindow
There will be a lot of non-shared functionality between a window for remote search plugin management and what I suspect local plugins will need as I want to add a download manager.
It will likely be easier to maintain to keep both entirely seperate, so this is a preperation for adding a LocalSearchPluginsWindow.