Now correctly uses RemoteSearchPluginService
It was still manually calling QBittorrentService bypassing the Singleton model leaving it out of the loop on updates.

Also slightly increased the task delay for `WaitForQbittorentPluginSyncAsync()` as sometimes it missed a beat.
Better yet would be to have it check / compare the list for changes and send requests until it does or times out... but that's a lot of work and there is a manual refresh button.