Changed how presets/saving are presented
Rather than a big bar at the bottom the save button is now part of the title section and has a flyout to access the other save/export options.
This frees up space to move the now missing presets to the View (instead of Preview) panel and have them render at scale.

Removed Mode switching code - now housed in parent View control as it contains the switching component