Switched to using index based system for history
Previously the code was using a lot of ElementAt and many updates in different places. Now it uses the index and monitors updates to the index.
Also added `AddToHistory` to manually trigger adding the current LogoColorsRecord to the history (intention is doing so when the colorpicker is closed rather than for every color being dragged over in it)