Prevented empty (no files) AddTorrentRequests from being send
Because the actions will lead to changes of .torrent files in the directory (renaming, deleting or moving means there'll be done) it leads to a change in .torrent files in the monitored directory so the service was triggering itself.
Now opts out early if there's no .torrent files and thus prevents empty requests from being send.